feat: 🎸 Move (back) to CommonJS
