+++ .eslintrc.yml
@@ -26,6 +26,7 @@ plugins:
   - import
   - prettier
   - '@typescript-eslint'
+  - 'chai-friendly'
 
 settings:
   import/parsers:
@@ -47,6 +48,7 @@ rules:
   '@typescript-eslint/no-explicit-any': off
   '@typescript-eslint/explicit-function-return-type': off
   '@typescript-eslint/no-unused-vars': warn
+  '@typescript-eslint/no-unused-expressions': off,
   '@typescript-eslint/no-unsafe-assignment': off
   '@typescript-eslint/no-unsafe-member-access': off
   '@typescript-eslint/no-unsafe-return': off
