+++ package.json
@@ -22,12 +22,11 @@
     "ignore-walk": "^3.0.4",
     "js-yaml": "^4.1.0",
     "lodash-es": "^4.17.21",
-    "mocha": "^9.1.0",
     "node-fetch": "^2.6.1",
     "tar": "^6.1.0",
     "validator": "^13.6.0",
     "yargs": "^17.0.1",
-    "zx": "github:google/zx"
+    "zx": "^4.2.0"
   },
   "description": "Otomi Core is an opinionated stack of Kubernetes apps and configurations. Part of Otomi Container Platform.",
   "devDependencies": {
@@ -35,17 +34,15 @@
     "@commitlint/config-conventional": "^11.0.0",
     "@hkdobrev/run-if-changed": "^0.3.1",
     "@types/async-retry": "^1.4.3",
-    "@types/chai": "^4.2.16",
-    "@types/chai-as-promised": "^7.1.3",
     "@types/debug": "^4.1.5",
     "@types/express": "^4.17.11",
     "@types/ignore-walk": "^3.0.2",
     "@types/js-yaml": "^4.0.1",
     "@types/lodash": "^4.14.168",
     "@types/lodash-es": "^4.17.4",
-    "@types/mocha": "^7.0.2",
+    "@types/mocha": "^9.0.0",
     "@types/node": "^16.7.1",
-    "@types/sinon": "^9.0.11",
+    "@types/sinon": "^10.0.4",
     "@types/sinon-chai": "^3.2.5",
     "@types/supertest": "^2.0.11",
     "@types/tar": "^4.0.4",
@@ -53,7 +50,8 @@
     "@types/yargs": "^17.0.0",
     "@typescript-eslint/eslint-plugin": "^4.28.0",
     "@typescript-eslint/parser": "^4.28.0",
-    "@vercel/ncc": "^0.28.6",
+    "chai": "^4.3.4",
+    "chai-as-promised": "^7.1.1",
     "commitizen": "^4.2.4",
     "cspell": "^4.2.8",
     "cz-conventional-changelog": "^3.3.0",
@@ -65,14 +63,19 @@
     "eslint-plugin-chai-friendly": "^0.7.2",
     "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-prettier": "^3.4.0",
+    "esm-loader-chaining-polyfill": "^0.1.1",
     "git-cz": "^4.7.6",
     "github-release-from-changelog": "^2.1.1",
     "husky": "^7.0.0",
     "json-schema-to-typescript": "^10.1.4",
     "json2ts": "^0.0.7",
     "lint-staged": "^10.5.4",
+    "mocha": "^9.1.2",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.1",
+    "run-if-changed": "^0.0.9",
+    "sinon": "^11.1.2",
+    "sinon-chai": "^3.7.0",
     "standard-version": "^9.3.0",
     "ts-node": "^10.1.0",
     "typescript": "^4.3.4"
@@ -123,9 +126,10 @@
     "spellcheck": "cspell '/docs/**/**.md' '/values-schema.yaml' '/*.md' '/.values/README.md'",
     "test": "TESTING=1 run-p lint validate-templates",
     "test:bats": "TESTING=1 binzx/otomi x bats -- -T bin/tests",
-    "test:ci": "TESTING=1 binzx/otomi x src/ci-tests.ts",
     "test:opa": "binzx/otomi x opa test policies -v",
     "test:ts": "mocha",
+    "test:cjs": "TS_NODE_COMPILER_OPTIONS='{\"module\":\"commonjs\"}' mocha",
+    "test:ts2": "npm run compile && mocha",
     "un-release": "sh -c 'git tag -d $1 && git push origin :refs/tags/$1' sh",
     "validate-templates": "TESTING=1 binzx/otomi validate-templates",
     "validate-values": "TESTING=1 binzx/otomi validate-values"
@@ -135,6 +139,6 @@
       "tag": true
     }
   },
-  "type": "module",
+  "type": "commonjs",
   "version": "0.14.10"
 }
