+++ src/cmd/destroy.ts
@@ -1,13 +1,13 @@
 import { existsSync, unlinkSync, writeFileSync } from 'fs'
 import { Argv } from 'yargs'
 import { $, nothrow } from 'zx'
-import { hf } from '../common/hf.js'
-import { cleanupHandler, prepareEnvironment } from '../common/setup.js'
-import { getFilename, getParsedArgs, logLevelString, OtomiDebugger, setParsedArgs, terminal } from '../common/utils.js'
-import { Arguments, helmOptions } from '../common/yargs-opts.js'
-import { ProcessOutputTrimmed, stream } from '../common/zx-enhance.js'
+import { hf } from '../common/hf'
+import { cleanupHandler, prepareEnvironment } from '../common/setup'
+import { getFilename, getParsedArgs, logLevelString, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
+import { Arguments, helmOptions } from '../common/yargs-opts'
+import { ProcessOutputTrimmed, stream } from '../common/zx-enhance'
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 const templateFile = '/tmp/otomi/destroy-template.yaml'
 const debug: OtomiDebugger = terminal(cmdName)
