+++ src/cmd/diff.ts
@@ -1,12 +1,12 @@
 import { Argv } from 'yargs'
-import { Arguments, decrypt } from '../common/crypt.js'
-import { hf } from '../common/hf.js'
-import { prepareEnvironment } from '../common/setup.js'
-import { getFilename, getParsedArgs, logLevelString, OtomiDebugger, setParsedArgs, terminal } from '../common/utils.js'
-import { helmOptions } from '../common/yargs-opts.js'
-import { ProcessOutputTrimmed } from '../common/zx-enhance.js'
+import { Arguments, decrypt } from '../common/crypt'
+import { hf } from '../common/hf'
+import { prepareEnvironment } from '../common/setup'
+import { getFilename, getParsedArgs, logLevelString, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
+import { helmOptions } from '../common/yargs-opts'
+import { ProcessOutputTrimmed } from '../common/zx-enhance'
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
 
 export const diff = async (): Promise<ProcessOutputTrimmed> => {
