+++ src/cmd/encrypt.ts
@@ -1,9 +1,9 @@
 import { Argv } from 'yargs'
-import { Arguments, encrypt as encryptFunc } from '../common/crypt.js'
-import { prepareEnvironment } from '../common/setup.js'
-import { getFilename, getParsedArgs, OtomiDebugger, setParsedArgs, terminal } from '../common/utils.js'
+import { Arguments, encrypt as encryptFunc } from '../common/crypt'
+import { prepareEnvironment } from '../common/setup'
+import { getFilename, getParsedArgs, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
 
 export const encrypt = async (): Promise<void> => {
