+++ src/cmd/gen-drone.ts
@@ -1,8 +1,8 @@
 import { writeFileSync } from 'fs'
 import { Argv } from 'yargs'
-import { env } from '../common/envalid.js'
-import { hfValues } from '../common/hf.js'
-import { getImageTag, prepareEnvironment } from '../common/setup.js'
+import { env } from '../common/envalid'
+import { hfValues } from '../common/hf'
+import { getImageTag, prepareEnvironment } from '../common/setup'
 import {
   BasicArguments,
   getFilename,
@@ -12,13 +12,13 @@ import {
   rootDir,
   setParsedArgs,
   terminal,
-} from '../common/utils.js'
+} from '../common/utils'
 
 export interface Arguments extends BasicArguments {
   dryRun?: boolean
 }
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
 
 export const genDrone = async (): Promise<void> => {
