+++ src/cmd/playground.ts
@@ -1,13 +1,13 @@
 import { Argv } from 'yargs'
-import { prepareEnvironment } from '../common/setup.js'
-import { BasicArguments, getFilename, OtomiDebugger, setParsedArgs, terminal } from '../common/utils.js'
+import { prepareEnvironment } from '../common/setup'
+import { BasicArguments, getFilename, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
 /**
  * This file is a scripting playground to test basic code
  * it's basically the same as EXAMPLE.ts
  * but loaded into the application to run.
  */
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
 
 export const playground = async (): Promise<void> => {
