+++ src/cmd/score-templates.ts
@@ -1,12 +1,12 @@
 import { existsSync, unlinkSync } from 'fs'
 import { Argv } from 'yargs'
 import { $, nothrow } from 'zx'
-import { hfTemplate } from '../common/hf.js'
-import { cleanupHandler, prepareEnvironment } from '../common/setup.js'
-import { getFilename, getParsedArgs, OtomiDebugger, setParsedArgs, terminal } from '../common/utils.js'
-import { Arguments, helmOptions } from '../common/yargs-opts.js'
+import { hfTemplate } from '../common/hf'
+import { cleanupHandler, prepareEnvironment } from '../common/setup'
+import { getFilename, getParsedArgs, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
+import { Arguments, helmOptions } from '../common/yargs-opts'
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 const templatePath = '/tmp/template.yaml'
 const debug: OtomiDebugger = terminal(cmdName)
