+++ src/cmd/status.ts
@@ -1,11 +1,11 @@
 import { Argv } from 'yargs'
 import { $ } from 'zx'
-import { prepareEnvironment } from '../common/setup.js'
-import { BasicArguments, getFilename, setParsedArgs } from '../common/utils.js'
+import { prepareEnvironment } from '../common/setup'
+import { BasicArguments, getFilename, setParsedArgs } from '../common/utils'
 
 type Arguments = BasicArguments
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 
 export const status = async (): Promise<void> => {
   const output = await $`helm list -A -a`
