+++ src/cmd/sync.ts
@@ -1,10 +1,10 @@
 import { Argv } from 'yargs'
-import { hf } from '../common/hf.js'
-import { prepareEnvironment } from '../common/setup.js'
-import { getFilename, getParsedArgs, logLevelString, OtomiDebugger, setParsedArgs, terminal } from '../common/utils.js'
-import { Arguments, helmOptions } from '../common/yargs-opts.js'
+import { hf } from '../common/hf'
+import { prepareEnvironment } from '../common/setup'
+import { getFilename, getParsedArgs, logLevelString, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
+import { Arguments, helmOptions } from '../common/yargs-opts'
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
 
 export const sync = async (): Promise<void> => {
