+++ src/cmd/validate-values.ts
@@ -1,9 +1,9 @@
 import Ajv, { DefinedError, ValidateFunction } from 'ajv'
-import { unset } from 'lodash-es'
+import { unset } from 'lodash'
 import { Argv } from 'yargs'
 import { chalk } from 'zx'
-import { hfValues } from '../common/hf.js'
-import { prepareEnvironment } from '../common/setup.js'
+import { hfValues } from '../common/hf'
+import { prepareEnvironment } from '../common/setup'
 import {
   getFilename,
   getParsedArgs,
@@ -12,10 +12,10 @@ import {
   rootDir,
   setParsedArgs,
   terminal,
-} from '../common/utils.js'
-import { Arguments, helmOptions } from '../common/yargs-opts.js'
+} from '../common/utils'
+import { Arguments, helmOptions } from '../common/yargs-opts'
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
 
 const internalPaths: string[] = ['apps', 'k8s', 'services', 'sops', 'teamConfig.services']
