+++ src/cmd/wait-for.ts
@@ -1,5 +1,5 @@
 import { Argv } from 'yargs'
-import { prepareEnvironment } from '../common/setup.js'
+import { prepareEnvironment } from '../common/setup'
 import {
   BasicArguments,
   getFilename,
@@ -8,9 +8,9 @@ import {
   setParsedArgs,
   terminal,
   waitTillAvailable,
-} from '../common/utils.js'
+} from '../common/utils'
 
-const cmdName = getFilename(import.meta.url)
+const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
 
 export interface Arguments extends BasicArguments {
