+++ src/common/crypt.ts
@@ -1,9 +1,9 @@
 import { EventEmitter } from 'events'
 import { existsSync, statSync, utimesSync, writeFileSync } from 'fs'
-import { chunk } from 'lodash-es'
+import { chunk } from 'lodash'
 import { $, cd, ProcessOutput } from 'zx'
-import { env } from './envalid.js'
-import { BasicArguments, OtomiDebugger, readdirRecurse, rootDir, terminal } from './utils.js'
+import { env } from './envalid'
+import { BasicArguments, OtomiDebugger, readdirRecurse, rootDir, terminal } from './utils'
 
 export interface Arguments extends BasicArguments {
   files?: string[]
