+++ src/common/setup.ts
@@ -1,14 +1,14 @@
 import { existsSync, readdirSync } from 'fs'
 import { fileURLToPath } from 'url'
 import { $, chalk } from 'zx'
-import { decrypt } from './crypt.js'
-import { env } from './envalid.js'
-import { hfValues } from './hf.js'
-import { BasicArguments, loadYaml, parser, terminal } from './utils.js'
-import { askYesNo } from './zx-enhance.js'
+import { decrypt } from './crypt'
+import { env } from './envalid'
+import { hfValues } from './hf'
+import { BasicArguments, loadYaml, parser, terminal } from './utils'
+import { askYesNo } from './zx-enhance'
 
 chalk.level = 2
-const dirname = fileURLToPath(import.meta.url)
+const dirname = __dirname
 
 let otomiImageTag: string
 let otomiClusterOwner: string
