+++ src/common/values.ts
@@ -2,9 +2,9 @@ import cleanDeep, { CleanOptions } from 'clean-deep'
 import { existsSync } from 'fs'
 import { writeFile } from 'fs/promises'
 import { dump } from 'js-yaml'
-import { cloneDeep, isEmpty, isEqual, merge, omit, pick } from 'lodash-es'
-import { env } from './envalid.js'
-import { extract, flattenObject, getValuesSchema, loadYaml, terminal } from './utils.js'
+import { cloneDeep, isEmpty, isEqual, merge, omit, pick } from 'lodash'
+import { env } from './envalid'
+import { extract, flattenObject, getValuesSchema, loadYaml, terminal } from './utils'
 
 const objectToString = (obj: Record<string, any>): string => {
   return isEmpty(obj) ? '' : dump(obj)
