+++ src/common/zx-enhance.ts
@@ -1,6 +1,6 @@
 import { chalk, ProcessOutput, ProcessPromise, question } from 'zx'
-import { env } from './envalid.js'
-import { DebugStream, getParsedArgs } from './utils.js'
+import { env } from './envalid'
+import { DebugStream, getParsedArgs } from './utils'
 
 const MAX_RETRIES_QUESTION = 3
